url household, resident

+++ src/modules/resident/resident.controller.js
@@ -1,11 +1,10 @@
 import { HttpStatus, UserRole } from "../../constant.js";
 import ErrorResponse from "../../utils/ErrorResponse.js";
 import SuccessResponse from "../../utils/SuccessResponse.js";
-import * as residentService from "./resident.service.js";
-import { getHouseholdByHouseholdCode } from "../household/household.service.js";
-const getAllResident = async (req, res) => {
+import * as residentService from "./services/resident.service.js";
+const getList = async (req, res) => {
    try {
-      const resident = await residentService.getAllResident();
+      const resident = await residentService.getList(req.query);
       return res.status(HttpStatus.OK).json(new SuccessResponse(resident));
    } catch (error) {
       return res
@@ -15,33 +14,28 @@ const getAllResident = async (req, res) => {
          );
    }
 };
-const getResidentByHouseholdCode = async (req, res) => {
-   try {
-      const resident = await residentService.getResidentByHouseholdCode(
-         req.params.householdCode
-      );
-      return res.status(HttpStatus.OK).json(new SuccessResponse(resident));
-   } catch (error) {
-      return res
-         .status(HttpStatus.INTERNAL_SERVER_ERROR)
-         .json(
-            new ErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, error.message)
-         );
-   }
-};
+// const getResidentByHouseholdCode = async (req, res) => {
+//    try {
+//       const resident = await residentService.getResidentByHouseholdCode(
+//          req.params.householdCode
+//       );
+//       return res.status(HttpStatus.OK).json(new SuccessResponse(resident));
+//    } catch (error) {
+//       return res
+//          .status(HttpStatus.INTERNAL_SERVER_ERROR)
+//          .json(
+//             new ErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, error.message)
+//          );
+//    }
+// };
 const createResident = async (req, res) => {
    try {
-      let household = await getHouseholdByHouseholdCode(
-         req.params.householdCode
-      );
-      if (!household) {
-         return res
-            .status(HttpStatus.NOT_FOUND)
-            .json(new ErrorResponse(HttpStatus.NOT_FOUND, "Not found"));
-      }
-      req.body.householdCode = req.params.householdCode;
-      let person = await residentService.createResident(req.body);
-      return res.status(HttpStatus.OK).json(new SuccessResponse(person));
+      let resident = await residentService.createResident({
+         ...req.body,
+         idNguoiTao: req.userId,
+         ngayTao: new Date().toString(),
+      });
+      return res.status(HttpStatus.OK).json(new SuccessResponse(resident));
    } catch (error) {
       return res
          .status(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -52,15 +46,26 @@ const createResident = async (req, res) => {
 };
 const deleteResidentByID = async (req, res) => {
    try {
-      let rowAffected = await residentService.deleteResidentByID(
-         req.params.residentID
+      let resident = await residentService.updateResidentByID(
+         req.params.residentID,
+         {
+            idNguoiXoa: req.userId,
+            ngayXoa: new Date().toString(),
+         }
       );
-      if (rowAffected > 0)
-         return res.status(HttpStatus.OK).json(new SuccessResponse("OK"));
-      else
-         return res
-            .status(HttpStatus.NOT_FOUND)
-            .json(new ErrorResponse(HttpStatus.NOT_FOUND, "Not found"));
+      if (resident) {
+         let rowAffected = await residentService.deleteResidentByID(
+            req.params.residentID
+         );
+         if (rowAffected > 0)
+            return res
+               .status(HttpStatus.OK)
+               .json(new SuccessResponse(req.params.residentID));
+         else
+            return res
+               .status(HttpStatus.NOT_FOUND)
+               .json(new ErrorResponse(HttpStatus.NOT_FOUND, "Not found"));
+      }
    } catch (error) {
       return res
          .status(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -74,7 +79,7 @@ const getResidentByID = async (req, res) => {
       const resident = await residentService.getResidentByID(
          req.params.residentID
       );
-      if (res)
+      if (resident)
          return res.status(HttpStatus.OK).json(new SuccessResponse(resident));
       else
          return res
@@ -90,21 +95,26 @@ const getResidentByID = async (req, res) => {
 };
 const updateResidentByID = async (req, res) => {
    try {
-      // kiểm tra xem tồn tại household đó không
-      let household = await getHouseholdByHouseholdCode(req.body.householdCode);
-
       // kiểm tra xem tồn tại resident đó không
-      let resident = await residentService.getResidentByID(req.body.id);
-      if (!household || !resident) {
+      let resident = await residentService.getResidentByID(
+         req.params.residentID
+      );
+      if (!resident) {
          return res
             .status(HttpStatus.NOT_FOUND)
             .json(new ErrorResponse(HttpStatus.NOT_FOUND, "Not found"));
       }
-      let rowAffected = await residentService.updateResidentByID(req.body);
+      let rowAffected = await residentService.updateResidentByID(
+         req.params.residentID,
+         req.body
+      );
       if (rowAffected > 0) {
+         let residentUpdate = await residentService.getResidentByID(
+            req.params.residentID
+         );
          return res
             .status(HttpStatus.OK)
-            .json(new SuccessResponse(rowAffected));
+            .json(new SuccessResponse(residentUpdate));
       } else {
          return res
             .status(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -124,10 +134,10 @@ const updateResidentByID = async (req, res) => {
    }
 };
 export {
-   getAllResident,
    createResident,
-   getResidentByHouseholdCode,
+   // getResidentByHouseholdCode,
    deleteResidentByID,
    getResidentByID,
    updateResidentByID,
+   getList,
 };
